C++: expose DBF as dbf() in task model
The old API avoided relying on return value optimization. With a
decent compiler, this is no longer needed or useful. Alas, rewriting
the entire API is too much for now.

Further, fix a corner-case issue in the DBF computation if unsigned
long is used. This completes the fix given in
e17645921697351bd968f034a85299c02332ad16, which overlooked the
unsigned long version.