Add a new model for CAN bus messages
CAN bus messages are similar to real-time tasks, but the formulas
to compute the interference, response time, and execution time
depend on the CAN bus properties, and differs from that of a task
executing on a processor.

In addition, the CAN bus message set also has support for replication
of one or more tasks.