Remove schedcat/util/quantor.py
This commit removes the quantor.py file from schedcat/util/.  This file
simply defined two functions, forall() and exists(), for testing if a
function evaluates to true for all or any element of an iterable.  This
can be achieved more efficiently using the Python built-in functions
all() and any(), so this commit changes all the SchedCAT code to take
this approach instead.