Use fractions.gcd instead of our own
The fractions module provides a gcd() function that doesn't require any
effort on our part (i.e. we don't have to have our own test code for
it).  It handles negative parameters a bit differently, returning
a negative value if the second parameter is negative, but for the
things SchedCAT uses greatest common denominators for this difference
never comes up.  This commit uses that gcd() function instead of our
own, and uses the lcm() function defined in schedcat.util.math in the
schedcat.generator.generator_emstada module instead of defining a new
one there.